makefile: remove BIN_ prefix
From RIOT bindist example README:
RIOT allows for creating a "binary distribution", which can be used to ship
proprietary, compiled objects in a way that makes it possible to re-link
them against a freshly compiled RIOT.

As the need is to build mcu-firmware and not RIOT, this prefix is useless.
Moreover mcu-firmware is not a proprietary code.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>