robotics/platform: add task task_start_shell()
This task waits the Enter key to be pressed and up a flag.
If the flag is not up before PF_START_COUNTDOWN seconds, the task thread
is stopped and it means the robot could start the game automatically.
If the flag is up by pressing Enter key, the robot does not start the game
and a RIOT shell is available.
Commands could be added to this shell for debug, test, calibration or
configuration purposes.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>