controllers: quadpid: correct anti-blocking
Anti blocking was not working because it was based on a bad algorithm.
The right test to perform is to check the current speed and the error.
The current must not be below a defined treshold and the speed error must
not be greater than a defined treshold defined in the platform.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>