platforms: cogip2019-cortex: CC1101 radio bring-up
This commit activate the CC1101 radio module in cortex 2019, including
SPI2 peripheral activation.

Warning: RIOT is limited to used a SPI IP which is muxed on same
alternate function in MCU pinmux. In our HW, the SPI2 signals are
not wired on the same gpio port and requires different alternate
function. So, we added a call to:

gpio_init_af(spi_config[bus].mosi_pin, GPIO_AF7);

after cc110x driver init. But, this is not sufficient and this should
also be done in the driver itself !

cc110x-spi.c, l.77:
(...)
    /* Switch MISO/GDO1 to spi mode */
    spi_init_pins(dev->params.spi);
    gpio_init_af(GPIO_PIN(PORT_C, 1), GPIO_AF7);//spi_config[bus].af);
}
(...)

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>