platforms: re-arrange makefiles
Common platform was defined as an application and not as a module.
Thus common Makefile configurations was still duplicated in each platforms.

Now each platforms can rely on common code by defining
MCUFIRMWARE_PLATFORM_BASE in its own Makefile.

Moreover, to improve readability cogip2019-cortex-common has been renamed
to cortex-common. No need to specify "cogip" as it is cogip repository and
the year is useless as git tags mentions it.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>