travis: use a non-rolling RIOT version
Whereas it can be useful to track upstream, using a moving branch can lead
to unrepredictable build behavior due to design updates in RIOT OS
not related to mcu-firmware on-going developments.

This update change the branch to clone, from master to a maintenance
branch aligned on a stable RIOT OS release.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>