platform: display calibration info message when built-in
This display informative message when calibration is built-in, even if
the ENABLE_LOG macro is set to 0.

Currently, when application is launched with calibration activated,
nothing happend, the time counter is not display, user need to look at
the code to know it should press a key.

This intend to avoid confusion when running this mode.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>