simu: pid: add CurveOverTime object
This introduce "CurveOverTime" object, which is a derivation of PlotItem
object class. Instances of this class are intended to display two datas
coming from the robot telemetry over time.

Typically, this will be used to display numerical data probed in the
control loop, especially 'command' data and their associated
'measurements'.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>