controllers: add a cycle counter incremented on each control loop turn
This introduce a cycle counter incremented each time the feedback
loop controller is trigged to perform control of the system.

This is required to compute some time dependant functions, especially
speed orders patterns (unit pulse, ramp function) which will be used
to identify the robot model and thus tune the PID coefficients.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>