simu: pid: add RobotTelemetry object
Introduce a new class object 'RobotTelemetry' intend to store multiple
measurements points of the control loop at a single instant in time.

Instances of this object are created each time the controller is
starting a new cycle_counter epoch, and stored in a thread-safe FIFO
handled by class methods of the object.

On the UI side, each time the update callback timer is trigged, it will
recover all those instances and update graphs according to the
measurements.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>