simu: pid: add GraphWindow object
Current design is using pg.GraphicsWindow instance as the main UI
object. This commit introduce a custom window class 'GraphWindow' which
is a derivation of QMainWindow, and directly embed the UI design in its
core.

Thus, all global implementation of main graph window is moved in this
dedicated object class. Basically, the static UI build and graph
update_plot.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>