simu: pid: do not launch native process in uart mode
The MCU firmware simulation binary (build for native target), is launch
whereas the script is used in simulation or on the real embedded system
through UART. This commit remove the subprocess start when the script is
used in UART mode.

Also, Thread identifier had been updated to reflect the characters
direction treated by the thread.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>