simu: pid: implement save as CSV action
To analyse robot movements and identify a transfer function, an
external tool such as GNU Octave will be used. This commit implement
an initial save to CSV feature to enable this kind of analysis.

CSV files are stored in a sub-directory 'csv_outputs' next to the
python script location, and named using a timestamp each time user
click on the icon in the toolbar.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>