ctrl: update get_speed_order api
ctrl_get_speed_order is returning a pointer on a vector structure.

This commit changes the API and allow the output vector to be passed
as a reference on a structure ; external to object context.

Thus, if required the user can then perform further computation on
the returned value, without modifying the object itself.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>