calib: quadpid: introduce model identification sequences
This commit add two command to characterize mobile robot movement
and help in model identification.
Both sequences, for linear and angular movements, are running with
no PID activated in the control loop. An impulse command is sent to
motors and encoders values can be probed to observe mechanical behavior
of the system.

Signed-off-by: Yannick Gicquel <ygicquel@gmail.com>