sys: tracefd: add trace file descriptor module
This module provides an API to write trace files on sdcard.

Since it may be difficult to determine when to close the files
to effectively flush data on disk, and to avoid opening/closing files
around each printf, this module starts a thread that
periodically closes/opens files to flush data.
This thread can be disabled to allow manual handling of opening/closing files.

Using the lock/unlock API, it is possible to execute
consecutive prints in thread safe manner.

The module relies on 'sdcard_spi' driver, so SPI parameters must be
provided in the board/platform/application.

Native boards are also supported, files being written in '/tmp'.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>