lib: avoidance/obstacles: add multi-form obstacles
Using oriented polygons leads to collision problems when the robot
rotates on itself, as the obstacle rotates too.
Moreover presuming of obstacle angle is hazardous as it is a perfect
supposition.
Thus the most secure modelisation of an obstacle is a circle.
However circles do not provide points for navigation.
To solve this, an AABB bounding box is added around the circle with
a small treshold to anticipate circle measurement imprecisions.
Corners of this bounding box became the available points for
navigation, keeping circles for collision detection only.
As this commit is modifying obstacle parameters, it also
uniformize all 'lidar_obstacles.c' as all platforms are using the
same lidar obstacle management.
However circles are more dedicated to dynamic obstacles.
Static obstacles should always be defined by polygons.
Thus modify 'obstacle_t' structure to add geometric obstacle 'form'
property as an union. Union type is selected through 'type' member.
This has a huge impact on all modules using or depending on 'obstacles'.
'collisions' and 'avoidance' are mainly impacted as it is also the
opportunity to reorder and clean them.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>