lib: avoidance: handle all obstacles type
Avoidance should not care about the obstacle type.
Thus the avoidance get the obstacle radius which is in case of a
polygon the distance between the center and the most far point of
the polygon.
This radius is used to create the bounding box around the
obstacle to create avoidance intermediate poses.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>