boards: cogip-board: fix default UARTs DMA
DMA configuration is protected by 'MODULE_PERIPH_DMA' macro when
'periph_dma' module is a board feature.
Moreover, DMA stream and channel should be respectively set to
'DMA_STREAM_UNDEF' and 'UINT8_MAX' as DMA configuration is only
performed for LDS-01 driver at the time of this commit.
This driver configures stream and channel in a nearly hard coded way.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>