makefiles: add global_vars.mk.inc
This file contains all variables that need to be exported to become
global to all makefiles.
The export is thus done at the top makefile levels and should not be
done in modules Makefile.
This will be used by incoming embedded-proto package for the
variable PROTOBUF_PATH that will list all directories
containing .proto files for the built application.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>
Signed-off-by: Eric Courtois <eric.courtois@gmail.com>