lib: wizard: add new library
This module is used to communicate with the Copilot at the beginning
of a game to configure dynamic parameters or play a check-list.
It provides predefined Protobuf messages (bool, int, float,
str, list, ...) that can be sent using uartpb module.
After sending a message, the module waits for a response of the same type, with
its value updated.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>