lib: path: add unreachable() and next() method
With default Path class, unreachable() and next() just act like ++ and skip the
current point.
In more complex Path classes, it is usefull to notify the Path that the
current pose is unreachable or that we request a new position and let
it decide what to do next in each case.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>