sys: remove tracefd module
The tracefd module was made for UART communication with Raspberry Pi
in json format and for writing logs to SD-Card.
As both of them are not used anymore, tracefd becomes useless and must
be removed.
The printf and logf functions are replaced by COGIP_DEBUG_COUT and
COGIP_DEBUG_CERR macros that are simply printing log messages on
console. These macros are temporary and should be replaced in the
future by a logger module.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>