lib: path: add Protobuf message for cogip::path::Pose class
In the PathPose message, the speeds are defined as an enum (MIN, NORMAL
and MAX) because in mcu-firmware, the speeds (linear or angular) are
float-point values, but then can take only 3 values: low, normal and max
speed. These values depend on the platform, so on the Raspberry
side, we only need to define the speed levels instead of the real values.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>