sys: introduce thread module
The thread module is used to monitor threads loop and period overshots
per thread.
Instead of calling RIOT API directly to perform a periodic wakeup, each
module has to call thread_ztimer_periodic_wakeup() function to do so.
In case of period overshot, a warning message is displayed in the
console.
If the sysmon module is enabled, statistics are saved by the sysmon
module.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>