boards: cogip-board-ng: add the 3 motors of cup2023
The cup2023 application needs specific motor configuration for
actuators.
However the motors configuration is done in the generic file board.h.
This commit introduces specific code in a generic file because the RIOT
driver is done this way.
Modifying the RIOT driver should be done in a near future to handle such
specific cases.

As some motors do not have direction or enable pin but must be driven by
the same timer and the same API, use pin B15 as a fake output GPIO.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>