introduce cqfd
From cqfd README:
  cqfd provides a quick and convenient way to run commands in the
  current directory, but within a Docker container defined in a per-
  project config file.

Since Ubuntu 21.04, a bug[1] appeared leading to a lot of SIGFPE signals
killing the application when running on native architecture.
This bug is hard to solve and even to investigate. To solve this, use
cqfd to build and run the native app_test application in an Ubuntu 20.04
docker container based image.

Follow README for instructions.

[1]: https://github.com/RIOT-OS/RIOT/issues/495#issuecomment-1254664077

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>