platforms: pegasus: introduce LxMotor
The LxMotor class drives LX-15D servomotor in motor mode. The command is
a percentage of the maximum speed, signess of the command is the
direction.
As LxMotor and LxServo objects share the same UART, initialization has
to be done out of the pf_servos.cpp file and will be done in a further
commit in pf_actuators.cpp.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>