platforms: pegasus: disable actuators on emergency button press
When emergency button is pressed, 24V is off and then it sets down a
GPIO with triggers an interrupt.
On this interrupt, an event is generated to disable actuators and
protobuf message for actuators commands reception.
If emergency button is released, protobuf messages for actuators can be
sent again to the firmware.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>