motion_control: speed_filter: use current speed to check acceleration
Using speed_order means we do not care about current robot speed, so
limiting acceleration could be a non-sens.
Worst in case of brutal stop, as the previous speed order is stored, the
robot will move in an unwanted direction.
Acceleration should always be computed from current speed to ensure it
is computed from a real state.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>