motion_control/speed_filter: introduce anti-blocking
Check if the current speed of the robot is below a speed threshold,
typically set low.
Additionally, ensure the robot isn't accelerating by comparing the
current speed with the speed order from the previous cycle. If the
difference between these two speeds exceeds a threshold, it indicates
that the robot hasn't accelerated since its previous cycle.
Then increment a counter for cycles during which the robot is supposedly
blocked. This evaluation cannot be done within a single cycle, so the
counter is incremented.
After a certain number of cycles, the robot is considered blocked.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>