drive controller: add differential drive controller implementation
COGIP robots are differential. So, implement the differential drive
controller in order to convert robot polar speed into speed ratio for
each motors.

The speed ratio is defined using the motor electrical informations found
in the datasheet. This ratio reflects the percentage of nominal motor
voltage to apply in order to reach the velocity wanted.

Signed-off-by: Mathis Lécrivain <lecrivain.mathis@gmail.com>