boards/cogip-board: set mandatory fdcan loop delay
Compensate for transceiver loop delay in FDCAN timing

The transceiver loop delay introduces a lag between sending and
receiving messages in FDCAN due to internal processing. This delay
affects bit timing, especially in high-speed communication, and can lead
to synchronization issues.

This commit adjusts this timing parameters as it is mandatory for the
new RIOT fdcan driver implementation.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>