app: split app_conf in multiples headers for each robot configuration
As the motion-control application will be used for 1 robot and 4 small
robot (named PAMI in french robotic cup), it is convenient to define
one configuration header by robots.

By doing this, we can easily build the firmware for a dedicated robot by
using the following makefile target.

make ROBOT_ID=1

As the simulation considers the speed order as the current speed, the
entire motion control process is executed, but the QDEC simulation
overrides its values.
Thus, the robot will move perfectly regardless of how the motion control
parameters are set.

Signed-off-by: Mathis Lécrivain <lecrivain.mathis@gmail.com>
Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>