app: split app_conf in multiples headers for each robot configuration
As the motion-control application will be used for 1 robot and 4 small
robot (named PAMI in french robotic cup), it is convenient to define
one configuration header by robots.

By doing this, we can easily build the firmware for a dedicated robot by
using the following makefile target.

make ROBOT_ID=1

Signed-off-by: Mathis Lécrivain <lecrivain.mathis@gmail.com>