motion_control: refactor SpeedPIDController to compute error internally
Previously, the speed error had to be computed by a separate controller
and passed via IO keys. This created unnecessary coupling and required
an extra controller in the chain. By computing the error internally from
speed_order and current_speed, the controller becomes self-contained and
the pipeline simpler to configure.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>